Refs #256 - integrate with wsrep API changes to pass global TO seqno through replicaiton calls
In addition some refactoring and a lot of cleanups on wsdb was made:
- global seqno is now part of a write set structure
- added more return code checks and assertions
- got rid of ws_start_cb() as redundant now
- write set is now disposed off when all locks are released
- all seqno types are now signed integers


git-svn-id: https://devel.codership.com/svn/galera/trunk@1412 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ wsdb/src/dump_trxfile.c
@@ -264,7 +264,7 @@ static int dump_trx(struct block_info *bi) {
     rcode = get_data_from_block(bi, sizeof(struct trx_hdr), (void *)&trx); 
     if (rcode != WSDB_OK) return rcode;
 
-    fprintf(stdout, "TRX: %llu\n", (unsigned long long)trx.trx_seqno);
+    fprintf(stdout, "TRX: %lld\n", trx.trx_seqno);
 
     return (WSDB_OK);
 }
